Move Marker to core
